refactor: restructure package hierarchy
domain 및 parser 패키지 분리
LottoManager -> LottoController 클래스명 변경
