+++ src/main/java/lotto/domain/LottoConstants.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 public final class LottoConstants {
     public static final int MIN_LOTTO_NUMBER = 1;
