+++ src/main/java/lotto/parser/PurchaseAmountParser.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.parser;
 
 
 public class PurchaseAmountParser {
