+++ src/main/java/lotto/parser/WinningNumbersParser.java
@@ -1,4 +1,6 @@
-package lotto;
+package lotto.parser;
+
+import lotto.domain.LottoConstants;
 
 import java.util.Arrays;
 import java.util.Set;
