+++ src/test/java/lotto/BonusNumberParserTest.java
@@ -1,6 +1,6 @@
 package lotto;
 
-import org.junit.jupiter.api.BeforeEach;
+import lotto.parser.BonusNumberParser;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
