+++ src/test/java/lotto/PurchaseAmountParserTest.java
@@ -1,9 +1,8 @@
 package lotto;
 
+import lotto.parser.PurchaseAmountParser;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
